refactor(cloudflare-metrics): split graphql query builders from transport
the 418-line graphql-client.ts mixed the CloudflareGraphQLClient transport
(retries, counters, http) with ~150 lines of pure query-building helpers
(buildDatasetSelection, buildBatchedAccountQuery, buildBatchedZoneQuery,
buildFilterObject, groupErrorsByAlias, chunkArray, ACCOUNT_BATCH_CHUNK_SIZE).
pull the pure helpers into their own file so the transport layer isn't
responsible for string manipulation, and the builders are trivially
testable in isolation.

graphql-client.ts keeps re-exports so existing imports don't move.
