+++ apps/cloudflare-metrics/src/graphql-builders.ts
@@ -1,0 +1,186 @@
+import type { DatasetQuery, GraphQLResponse } from './types.js';
+
+/** Default row limit on every `*AdaptiveGroups` selection. */
+export const DEFAULT_DATASET_LIMIT = 9999;
+
+/**
+ * Cloudflare GraphQL rejects batched queries above ~50 aliased datasets
+ * with a "too many nodes, zones and accounts" error. Empirically we land
+ * safely at 25 per chunk — well under the ceiling, which leaves headroom
+ * for datasets with extra blocks/dimensions.
+ */
+export const ACCOUNT_BATCH_CHUNK_SIZE = 25;
+
+/**
+ * Split an array into fixed-size chunks. Pure helper used by the
+ * batched-fetch path in `graphql-client.ts`.
+ */
+export function chunkArray<T>(items: readonly T[], size: number): T[][] {
+  if (size <= 0) {
+    return items.length > 0 ? [[...items]] : [];
+  }
+  const chunks: T[][] = [];
+  for (let i = 0; i < items.length; i += size) {
+    chunks.push(items.slice(i, i + size) as T[]);
+  }
+  return chunks;
+}
+
+/**
+ * Emits the inner selection for a single dataset — the aliased field, its
+ * arguments, and the selection set — without the surrounding
+ * `viewer.accounts { ... }` wrapper. Used by the batched query builders so
+ * one query can carry many dataset selections.
+ */
+export function buildDatasetSelection(dataset: DatasetQuery, alias?: string): string {
+  const dimensionSelection = dataset.dimensions.join(' ');
+  const blockSelections = (Object.entries(dataset.blocks) as Array<[string, readonly string[] | undefined]>)
+    .filter(([, fields]) => fields && fields.length > 0)
+    .map(([block, fields]) => `${block} { ${(fields ?? []).join(' ')} }`)
+    .join(' ');
+  const topLevelSelection = (dataset.topLevelFields ?? []).join(' ');
+  const limit = dataset.limit ?? DEFAULT_DATASET_LIMIT;
+  const aliasPrefix = alias && alias !== dataset.field ? `${alias}: ` : '';
+  return `${aliasPrefix}${dataset.field}(limit: ${limit}, filter: $filter${orderByClause(dataset)}) {
+        dimensions { ${dimensionSelection} }
+        ${topLevelSelection}
+        ${blockSelections}
+      }`;
+}
+
+const SCHEDULED_INVOCATIONS_SELECTION = `workers_scheduled: workersInvocationsScheduled(limit: ${DEFAULT_DATASET_LIMIT}, filter: $filter) {
+        scriptName
+        cron
+        status
+        datetime
+        cpuTimeUs
+      }`;
+
+/**
+ * Builds a single GraphQL query that pulls many account-scope datasets in
+ * one HTTP request by aliasing each dataset. All datasets in the batch must
+ * share the same filter granularity (datetime or date) because they share
+ * the `$filter` variable.
+ */
+export function buildBatchedAccountQuery(
+  datasets: readonly DatasetQuery[],
+  includeScheduledInvocations = false,
+): string {
+  const selections = datasets.map((d) => buildDatasetSelection(d, d.key));
+  if (includeScheduledInvocations) {
+    selections.push(SCHEDULED_INVOCATIONS_SELECTION);
+  }
+  return `query CloudflareMetricsAccountBatch($accountTag: String!, $filter: JSON!) {
+  viewer {
+    accounts(filter: { accountTag: $accountTag }) {
+      ${selections.join('\n      ')}
+    }
+  }
+}`;
+}
+
+/**
+ * Builds a single GraphQL query that fetches one dataset across many
+ * zones. Each zoneTag becomes its own aliased `zones(...)` block.
+ */
+export function buildBatchedZoneQuery(zoneTags: readonly string[], dataset: DatasetQuery): string {
+  const safeZoneTags = zoneTags.map((tag) => {
+    if (!/^[a-zA-Z0-9_-]+$/.test(tag)) {
+      throw new Error(`Invalid zoneTag for batched query: ${tag}`);
+    }
+    return tag;
+  });
+  const selections = safeZoneTags
+    .map(
+      (tag, index) => `    z${index}: zones(filter: { zoneTag: "${tag}" }) {
+      ${buildDatasetSelection(dataset, dataset.field)}
+    }`,
+    )
+    .join('\n');
+  return `query CloudflareMetricsZoneBatch($filter: JSON!) {
+  viewer {
+${selections}
+  }
+}`;
+}
+
+/**
+ * Builds the filter JSON object passed via `$filter` in both batched and
+ * unbatched queries. Picks date-vs-datetime fields based on the dataset's
+ * declared filter granularity; all datasets in a batch must agree on this.
+ */
+export function buildFilterObject(
+  dataset: DatasetQuery | null,
+  range: { start: Date; end: Date },
+): Record<string, unknown> {
+  const filter: Record<string, unknown> = dataset?.extraFilter ? { ...dataset.extraFilter } : {};
+  if ((dataset?.filterGranularity ?? 'datetime') === 'date') {
+    filter.date_geq = formatDateOnly(range.start);
+    filter.date_leq = formatDateOnly(range.end);
+  } else {
+    filter.datetime_geq = range.start.toISOString();
+    filter.datetime_leq = range.end.toISOString();
+  }
+  return filter;
+}
+
+/**
+ * Converts the GraphQL `errors[]` array into a map from alias → message.
+ * Cloudflare's error payloads include a `path` that begins with the
+ * viewer → accounts/zones → aliased field; we walk it until we find an
+ * alias we recognise. Errors without a resolvable alias are grouped under
+ * an empty key and handled by the caller.
+ */
+export function groupErrorsByAlias(
+  errors: GraphQLResponse<unknown>['errors'] | null | undefined,
+): Record<string, string> {
+  const result: Record<string, string> = {};
+  if (!errors) {
+    return result;
+  }
+  for (const error of errors) {
+    const alias = findAliasInPath(error.path);
+    if (alias) {
+      const existing = result[alias];
+      result[alias] = existing ? `${existing}; ${error.message}` : error.message;
+    }
+  }
+  return result;
+}
+
+function findAliasInPath(path: readonly (string | number)[] | undefined): string | undefined {
+  if (!path) {
+    return undefined;
+  }
+  // Typical path shape:
+  //   ['viewer', 'accounts', 0, '<alias>']
+  //   ['viewer', '<alias>', 0, '<innerField>']   (zone batches)
+  // Return the first path segment after a list index that isn't part of
+  // the `dimensions` / `sum` / etc. internals.
+  for (let i = 0; i < path.length; i++) {
+    const segment = path[i];
+    if (typeof segment !== 'string') {
+      continue;
+    }
+    if (segment === 'viewer' || segment === 'accounts' || segment === 'zones') {
+      continue;
+    }
+    return segment;
+  }
+  return undefined;
+}
+
+function orderByClause(dataset: DatasetQuery): string {
+  const timestampDim = dataset.timestampDimension ?? 'datetimeMinute';
+  // Grouping implicitly takes place on the dimensions, so orderBy helps
+  // make sure we get the most recent buckets within the limit.
+  if (dataset.dimensions.includes(timestampDim)) {
+    return `, orderBy: [${timestampDim}_ASC]`;
+  }
+  return '';
+}
+
+function formatDateOnly(date: Date): string {
+  // YYYY-MM-DD
+  return date.toISOString().slice(0, 10);
+}
